spapr_vscsi: fix data corruption on transfers above 128 KiB
scsi-disk hands the HBA at most SCSI_DMA_BUF_SIZE (128 KiB) per
vscsi_transfer_data() callback, so a single SCSI command carrying more
than that arrives in several chunks.  vscsi_srp_direct_data() remembers
how far into the descriptor it already is in req->cur_desc_offset -- but
that field is uint16_t, so the very first 128 KiB chunk does

    req->cur_desc_offset += 0x20000;      /* 0x20000 & 0xFFFF == 0 */

and wraps straight back to zero.  Every following chunk of the same
command is then fetched with buf_offset 0 and DMA'd over the START of
the guest buffer: the tail of the transfer overwrites the head, and the
rest of the buffer is never written at all.  The command still completes
with status 0 and no sense data, so the guest cannot notice.

Linux never hits this.  Its ibmvscsi initiator uses INDIRECT descriptors
for anything multi-page, and the indirect path resets cur_desc_offset to
0 at every descriptor boundary, so the counter never reaches 65536.  A
guest that describes the whole transfer with a single DIRECT descriptor
and issues 256 KiB commands does hit it on the first command.

Measured with a READ_10 of 2098 sectors, traced with
spapr_vscsi_transfer_data and spapr_vscsi_fetch_desc_done:

    queue_cmd      tag 0x2 CMD 0x28=READ_10 LUN 0 ret: 262144
    transfer_data  tag=0x2 len=0x20000
    fetch_desc     cur=0 offs=0x0 ret { va=0x84d28 len=0x40000 }
    transfer_data  tag=0x2 len=0x20000
    fetch_desc     cur=0 offs=0x0 ret { va=0x84d28 len=0x40000 }
    cmd_complete   tag=0x2 status=0x0

Same tag, same request, same target address twice.  The guest buffer
ends up holding the second half of the transfer at offset 0 and zeroes
above it.

struct srp_direct_buf.len is already uint32_t and vscsi_fetch_desc()
already takes buf_offset as unsigned, so widening the field is all that
is required.  The VMSTATE entries for cur_desc_num/cur_desc_offset are
commented out upstream, so the migration format is unaffected.

Fixes: 8ca8a17c9a16 ("pseries: rework PAPR virtual SCSI")
Cc: qemu-stable@nongnu.org
Message-Id: <20260820122644.19691-2-bernard.ladenthin@gmail.com>

+++ hw/scsi/spapr_vscsi.c
@@ -86,7 +86,14 @@ typedef struct vscsi_req {
     uint16_t                total_desc;
     uint16_t                cdb_offset;
     uint16_t                cur_desc_num;
-    uint16_t                cur_desc_offset;
+    /*
+     * Byte offset reached inside the current descriptor.  Must be at
+     * least 32 bit: scsi-disk delivers a command in SCSI_DMA_BUF_SIZE
+     * (128 KiB) chunks, so a DIRECT descriptor larger than 64 KiB would
+     * wrap a uint16_t back to zero and make every chunk after the first
+     * overwrite the head of the guest buffer.
+     */
+    uint32_t                cur_desc_offset;
 } vscsi_req;
 
 #define TYPE_VIO_SPAPR_VSCSI_DEVICE "spapr-vscsi"
